Removed a warning (from buildbot output of builder moab-32bit) for Tqdcfr.cpp. This -Wsign-compare warning is only reproducible in a 32-bit moab build.

+++ Tqdcfr.cpp
@@ -2329,7 +2329,7 @@ ErrorCode Tqdcfr::read_acis_records( const char* sat_filename )
       
         // get next occurrence of '#' (record terminator)
       ret = strchr(&(char_buf[buf_pos]), '#');
-      while (ret && ret+1-&char_buf[0] < bytes_left && *(ret+1) != '\n')
+      while (ret && (unsigned int)(ret+1-&char_buf[0]) < bytes_left && *(ret+1) != '\n')
         ret = strchr(ret+1, '#');
       if (NULL != ret) {
           // grab the string (inclusive of the record terminator and the line feed) and complete the record
